% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggyoudraw-package.R
\docType{package}
\name{ggyoudraw-package}
\alias{ggyoudraw}
\alias{ggyoudraw-package}
\title{ggyoudraw: 'You Draw It' Charts for 'ggplot2'}
\description{
Lets readers finish a line chart themselves before the real data are revealed, in the style of the 'You Draw It' charts of The New York Times. Provides a 'ggplot2' layer, geom_you_draw_line(), and you_draw_it(), which turns the plot into an interactive 'htmlwidget' built on 'ggiraph'. The widget works in the viewer, in 'R Markdown' and 'Quarto' documents and in 'shiny' apps, where the reader's guesses come back as an input.
}
\author{
\strong{Maintainer}: Arne Coutteau \email{c_arne@hotmail.com} [copyright holder]

}
\keyword{internal}
